+++ Server/Data/Models/TopExpenseCategory.cs
@@ -4,9 +4,9 @@ public class TopExpenseCategory
 {
     public required string Category { get; set; }
     
-    public double Expenditure{ get; set; }
+    public decimal Expenditure{ get; set; }
     
-    public double Total { get; set; }
+    public decimal Total { get; set; }
     
     public required string IconUrl { get; set; }
 }
